fix(tests): import real implementations instead of inline copies
Extract pure functions to dedicated modules so tests can import real
implementations instead of copy-pasting the logic.

- src/lib/api/scope-filter.ts: buildScopeFilter + canAccessScopes
- src/lib/wiki-utils.ts: slugify + parseTagInput + normalizeTagInputs

auth.ts and wiki.ts re-export from the new modules, maintaining
backward compatibility for existing consumers.

Fixes the inline-copy maintenance trap identified in PR #166 review:
tests that copy-paste implementation code rather than importing the real
functions will silently diverge when the source changes.

Refs: #166
