+++ src/__tests__/api/scope-filter.test.ts
@@ -1,50 +1,9 @@
 import assert from "node:assert/strict";
 import test, { describe } from "node:test";
-
-/**
- * Tests for buildScopeFilter and canAccessScopes from src/lib/api/auth.ts.
- *
- * These functions are pure but live in a module that transitively imports prisma,
- * so we inline the logic here to test without a database. This mirrors the
- * actual implementation and catches regressions in the scope-filtering logic.
- */
-
-// ─── Inline copies of the pure functions for isolated testing ───────────────
-// Keep these in sync with src/lib/api/auth.ts
-
-function buildScopeFilter(
-  allowedScopes: string[] | undefined,
-  extraWhere: Record<string, unknown> = {},
-): Record<string, unknown> {
-  if (allowedScopes?.includes("*")) {
-    return extraWhere;
-  }
-  if (!allowedScopes || allowedScopes.length === 0) {
-    return {
-      ...extraWhere,
-      restrictedTags: { isEmpty: true },
-    };
-  }
-  return {
-    ...extraWhere,
-    OR: [
-      { restrictedTags: { isEmpty: true } },
-      { restrictedTags: { hasSome: allowedScopes } },
-    ],
-  };
-}
-
-function canAccessScopes(
-  articleScopes: string[],
-  allowedScopes: string[] | undefined,
-): boolean {
-  if (articleScopes.length === 0) return true;
-  if (allowedScopes?.includes("*")) return true;
-  if (!allowedScopes || allowedScopes.length === 0) return false;
-  return articleScopes.some((s) => allowedScopes.includes(s));
-}
-
-// ─── buildScopeFilter ──────────────────────────────────────────────────────
+import {
+  buildScopeFilter,
+  canAccessScopes,
+} from "@/lib/api/scope-filter";
 
 describe("buildScopeFilter", () => {
   test("admin scope (*) returns only extraWhere", () => {
@@ -96,8 +55,6 @@ describe("buildScopeFilter", () => {
     });
   });
 });
-
-// ─── canAccessScopes ────────────────────────────────────────────────────────
 
 describe("canAccessScopes", () => {
   test("unrestricted articles are always accessible", () => {
