+++ src/lib/wiki.ts
@@ -7,50 +7,10 @@ import {
   buildSearchableCTE,
   buildSearchTsQuery,
 } from "@/lib/memory/article-search";
+import { normalizeTagInputs } from "./wiki-utils";
 
-export function slugify(text: string): string {
-  return (
-    text
-      .toLowerCase()
-      .replace(/[^a-z0-9 -]/g, "")
-      .replace(/\s+/g, "-")
-      .replace(/-+/g, "-")
-      .replace(/^-+|-+$/g, "")
-      .trim() || "untitled"
-  );
-}
-
-export function parseTagInput(raw: string | null | undefined): string[] {
-  if (!raw) return [];
-
-  return Array.from(
-    new Set(
-      raw
-        .split(",")
-        .map((tag) => tag.trim())
-        .filter(Boolean)
-    )
-  );
-}
-
-export interface NormalizedTagInput {
-  name: string;
-  slug: string;
-}
-
-export function normalizeTagInputs(tagNames: string[]): NormalizedTagInput[] {
-  const bySlug = new Map<string, NormalizedTagInput>();
-
-  for (const name of tagNames) {
-    const trimmed = name.trim();
-    const slug = slugify(trimmed);
-    if (!slug || bySlug.has(slug)) continue;
-
-    bySlug.set(slug, { name: trimmed, slug });
-  }
-
-  return Array.from(bySlug.values());
-}
+export { slugify, parseTagInput, normalizeTagInputs } from "./wiki-utils";
+export type { NormalizedTagInput } from "./wiki-utils";
 
 export async function buildTagConnections(tagNames: string[]) {
   if (!tagNames.length) return [];
